+++ lib/ex_icon/source.ex
@@ -162,7 +162,7 @@ defmodule ExIcon.Source do
       ssl: [
         verify: :verify_peer,
         cacerts: :public_key.cacerts_get(),
-        depth: 2,
+        depth: 3,
         customize_hostname_check: [
           match_fun: :public_key.pkix_verify_hostname_match_fun(:https)
         ]