+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Use canonical casing for SVG element and attribute names.
+- Validate `module_name` in configuration.
 
 ## [0.5.0] - 2026-08-02